requirements.txt
cmake==3.29.0.1
dlib==19.22.0
firebase-admin==6.5.0
gspread==6.1.2
Pillow==8.4.0
protobuf==4.25.3
matplotlib==3.5.2
numpy==1.20.0
opencv-python==4.9.0.80
pandas==0.25.3
scikit-learn==1.3.2
scipy==1.8.1


